move finish_upload() to api_v4.py
